// ui/frontend/src/visualizer/types.tsx
export type ElementProperties = {
  [pName: string]: string;
};

export type GraphNodeData = {
  id: string;
  turing_id: string;
  type: "selected" | "neighbor";
  node_type_name: string;
  properties: ElementProperties;
  neighborNodeIds?: string[];
  ins?: string[],
  outs?: string[],
};

export type GraphEdgeData = {
  id: string;
  turing_id: string;
  type: "connecting" | "neighbor";
  turing_source_id: string;
  turing_target_id: string;
  source_id?: string;
  target_id?: string;
  edge_type_name: string;
  properties: ElementProperties;
};

export type GraphNode = {
  group: "nodes";
  data: GraphNodeData;
};